cursor empty fill and no blink when not focused
When several kitty terminals are used, the blink happens in all of them
and the cursor appears filled as if active, it is not intuitive which
terminal is actually expecting user input and which one is not.
Terminals such as xfce4-terminal provide a non-blinking non-filled
cursor to show the user the cursor location while at the same time
showing it is not expecting user input (keyboard is active somewhere
else).