Mouse selection based on edges
Base mouse selection on the edges between cells rather than the cells
themselves.

For finding the start and end line in regular selection, the vertical
cell index is used rather than the edge.  However, the vertical edge
is used in rectangular selection mode.

Now allows selection of a single character.

This changes the semantics of screen_start_selection and
screen_update_selection to accept an edge index instead of cell index
for x, and an edge or cell index for y depending on whether
rectangular selection is enabled.

[asmadeus@codewreck.org: fixed rebase conflicts]
Closes #945