Fix hiding the window title in macOS
In b5c21632383f9ff27c13991c91c5a72187a1196a I introduced a bug where only the window title of the first OS window would be hidden. This commit fixes that bug.
The bug was caused by the code hiding the window title being in an `if` statement that only executes once.