make canberra sound theme configurable
On Linux, kitty uses libcanberra to play an audible bell following the
XDG Sound Theming Specification. The library expects the application to
include the sound theme in the context, otherwise it plays from the
"freedesktop" theme; it does not query the environment or configuration.

Instead of linking against Gtk and querying GtkSettings, I've chosen to
allow the sound theme to be configured with a new kitty setting
"bell_sound_theme".

Change-Id: Ib6168fdd4a9ed5a8fb635ba02340613fb35ac426