Premultiply resizing text with the alpha value
When the window is transparent, the resizing text needs to be premultiplied or the anti-aliasing will look wrong. The issue seems to be more pronounced on grey to white backgrounds.
To reproduce the issue start kitty with `kitty --config NONE -o background_opacity=0.3 -o resize_draw_strategy=size -o foreground=#899` and resize the window.