Update border width and padding for relayout
Fix for #2346: whenever relayout() is called, which happens when moving
between monitors (among other things), recompute the various window
padding/margin widths and border width and reset the current layout.
This ensures that these measurements look OK after moving between
monitors with different resolution.

(As an optimization perhaps we could check if the resolution actually
changed and do nothing if it didn't change, but maybe that's unnecessary
since this doesn't appear to be a a performance-critical spot.)