Do not overwrite EDITOR variable
The current startup logic will overwrite the EDITOR variable in the case
that both EDITOR and VISUAL are defined. This is because get_editor_from_env
checks VISUAL before EDITOR, but setup_environment overwrites EDITOR in
either case.

This change simply adds a check to only define the EDITOR variable in
the case that it's not already set.