Downgrade libstartup-notification-1.so non-existence to warning
I'm not sure how common this library actually is. On my system, an Arch
Linux system with 2,170 packages installed, including KDE (Plasma), I
didn't have it.

Googling around shows me that its purpose is just to show a loading
cursor when an application is loading. That sounds pretty Windows 98 to
me. I see how it could be useful, but certainly it's not an error not to
have this around.