Ensure resize when toggling between maximized window and fullscreen
When toggling between maximized window and traditional fullscreen, we
need to manually trigger windowDidResize to ensure that the fullscreen
window is resized correctly.

macOS probably ignores the change because the window size actually
stays the same. However, the disappearance of the title bar causes the
usable height to change, leading to the text being stretched when
switching to fullscreen, and this results in ugly artifacts.