feat: Use linear gamma for cell shaders
This patch converts the cell shaders to use linear gamma for all colors,
finally converting it back to sRGB at the end of the fragment shader.

To improve performance it uses a lookup-table to convert the 256-color input
values to the linear representation of their sRGB colors in the vertex shader,
and it also has branch-less implementations for linear-sRGB and sRGB-linear
conversions.