Allow sending notification when a command finishes
When the option `notify_on_command_finish` is set, and a command
takes more than `notify_on_command_finish_min_duration` seconds
to execute, then when it finishes, send a desktop notification.

The value of `notify_on_command_finish` can be:
never | unfocused | invisible | always
the latter three are interpreted the same way as in the `o=` option
of the desktop notification protocol.