Implement a simple scroll progress indicator
Shows a simple bar on the right edge of the window that moves up as you
scroll further back. There are apparently a lot of people that dont use
a pager for browsing large scrollbacks. I will never understand this,
but, what the hell I was in that code area anyway for other reasons.

TODO: Maybe make it a rounded rectangle