Prevent Fn key from scrolling to the bottom
When the Fn key is pressed it should not cause us to scroll to the bottom
of the scrollback. This is because Fn may be used to access movement keys
(e.g. on a MacBook keyboard, Fn+Up = Page Up).

Most keyboards do not expose Fn to the operating system as a separate key
event, but there are two known exceptions: Macs running Linux (generates
XF86Fn) and some ThinkPads (generates XF86WakeUp). Ignore both key
events when deciding whether to scroll to the bottom. For consistency,
do the same when deciding whether key events should hide the mouse.