Allow shell integration scripts to request prompt click events
This is done via the `prompt-click-events` keyword in
`$KITTY_SHELL_INTEGRATION`.

Also adds the `$KITTY_SUPPORTS_PROMPT_CLICK_EVENTS` environment
variable, which scripts can use to check for support before requesting
click events.

Note that the fish integration script is intentionally not modified, as
the fish integration seems to be built-in in newer versions.