tests: apply `os.utime` to created symlinks too
Nanosecond precision is not universally available on all platforms. For
example, nixpkgs python is built using Apple SDK 10.12 that does not
have `utimensat`. Because of that, python `os.stat` falls back to
microsecond precision.

This patch applies `os.utime` with zeroed nanoseconds to symlinks
created by the test case, the same way as it is already applied to other
files tested for transfer.