[POC] Blinking text (SGR 5/25) in kitty
Closes kovidgoyal/kitty#540.

This PR is not yet ready to be merged, and may never be depending on
@kovidgoyal's feedback given his skepticism in the original issue (and
my own, to be fair); critically, I have not made it optional yet.

I wrote this for my own personal use. Because most TE's ironically
ignore SGR 5/25 due to its lack of support, I knew I could rely on it to
really draw my attention in scripts, without giving reallt any other
application the ability to blink as there's no check for it.

I wanted to just leave this patch in kovidgoyal/kitty#540. But, I
couldn't comment, despite having contributed PR's in past. Kovid said
over email to post this PR as an alternative.