Fix hint label collision
When the number of hint labels exceeds the size of the hint alphabet,
some of the labels must be expanded to two or more characters.  However,
as currently implemented, it is possible for some longer hint labels to
start with the same text as a shorter label.

For example, if the hint label choices are "j" "k" "ja" "jb" and "jc",
then it is not possible for the user to select the mark associated
with "j" because the hint engine will still be trying to match
against "ja" or "jb" etc.

This commit ensures that longer hint labels never start with the
same text as any shorter hint labels.